Code

+++ GULEVICH_OSISP_prj_2024/code/blockPanel.h
@@ -8,11 +8,13 @@
 #define HEIGHT_BLOCK 2  //Высота окна
 
 typedef struct block{   //Структура о блочной инфе
-    int offset;         //Смещение
+    off_t offset;         //Смещение
     int block;          //Номер блока
     int block_size;     //Размер блока
     int sector;         //Номер сектора в блоке
     int sector_size;
 }block;
+
+//extern block blockStats;
 
 static void print_block(WINDOW *block_win, block block_stats);  //Вывод панельки
